requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.3.2
pymongo==4.6.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
Pillow==10.1.0
aiofiles==23.2.1